implemented payment entity and changed dc.migration to match the rest of the entities

+++ src/main/resources/db/migration/V2__create_orders.sql
@@ -1,18 +1,18 @@
 CREATE TABLE orders (
-    id int primary key,
-    customer_id int not null,
+    id bigint primary key,
+    customer_id bigint not null,
     status varchar(255) not null,
     currency varchar(3) not null,
-    total_amount decimal(10,2) not null,
+    total_amount decimal(19,2) not null,
     constraint fk_orders_customer foreign key (customer_id) references customers(id)
 );
 
 CREATE TABLE order_lines (
-    id int primary key,
-    order_id int not null,
-    product_id int not null,
+    id bigint primary key,
+    order_id bigint not null,
+    product_id bigint not null,
     quantity int not null,
-    price_amount decimal(10,2) not null,
+    price_amount decimal(19,2) not null,
     price_currency varchar(3) not null,
     constraint fk_order_lines_order foreign key (order_id) references orders(id),
     constraint fk_order_lines_product foreign key (product_id) references products(id)
